Moved enums to dedicated class
