+++ src/cli.rs
@@ -1,5 +1,5 @@
-use crate::utils::logging::LogFormat;
-use clap::{Parser, Subcommand};
+use crate::utils::enums::{Commands, LogFormat};
+use clap::Parser;
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
@@ -18,47 +18,6 @@ pub struct Args {
 
     #[command(subcommand)]
     pub command: Commands,
-}
-
-#[derive(Subcommand, Debug)]
-pub enum Commands {
-    /// Get information about Kubernetes resources
-    Get {
-        #[command(subcommand)]
-        resource: GetResource,
-    },
-}
-
-#[derive(Subcommand, Debug)]
-pub enum GetResource {
-    /// List pod images and their registries
-    Images {
-        /// Kubernetes namespace to query (defaults to "default", ignored when --node is specified)
-        #[arg(
-            short,
-            long,
-            default_value = "default",
-            conflicts_with = "all_namespaces"
-        )]
-        namespace: String,
-
-        #[arg(short = 'N', long = "node", conflicts_with = "all_namespaces")]
-        node: Option<String>,
-
-        #[arg(short, long)]
-        pod: Option<String>,
-
-        /// Filter pods by container image registry
-        #[arg(short = 'R', long = "registry")]
-        registry: Option<String>,
-
-        #[arg(short = 'A', long = "all-namespaces", conflicts_with = "namespace")]
-        all_namespaces: bool,
-
-        /// Output format (default: normal, wide: shows additional columns)
-        #[arg(short = 'o', long = "output", default_value = "normal", value_parser = ["normal", "wide"])]
-        output: String,
-    },
 }
 
 impl Args {
